refactor(demo): 拆分 MarkdownInputField 综合示例为独立单点 demo
